Remove unused functions in statistics_utils.py
This removes the unused functions and their tests. The tests then
remaining in `test_statistics_utils.py` have been moved to `test_job.py`
since they relate to the Job model and not to the utils functions.
