+++ tests/app/models/test_job.py
@@ -24,3 +24,35 @@ def test_still_processing(notify_admin, job_status, num_notifications_created, e
     )
     job = Job(json)
     assert job.still_processing == expected_still_processing
+
+
+@pytest.mark.parametrize(
+    "failed, delivered, expected_failure_rate", [(0, 0, 0), (0, 1, 0), (1, 1, 50), (1, 0, 100), (1, 4, 20)]
+)
+def test_add_rate_to_job_calculates_rate(failed, delivered, expected_failure_rate):
+    resp = Job(
+        {
+            "statistics": [
+                {"status": "failed", "count": failed},
+                {"status": "delivered", "count": delivered},
+            ],
+            "id": "foo",
+        }
+    )
+
+    assert resp.failure_rate == expected_failure_rate
+
+
+def test_add_rate_to_job_preserves_initial_fields():
+    resp = Job(
+        {
+            "statistics": [
+                {"status": "failed", "count": 0},
+                {"status": "delivered", "count": 0},
+            ],
+            "id": "foo",
+        }
+    )
+
+    assert resp.notifications_failed == resp.notifications_delivered == resp.failure_rate == 0
+    assert resp.id == "foo"
